Add API tests
This commit adds API tests, and makes some modifications
to already existing logic:

- Add an export called closeConnection to utils.js to allow
closing thedatabase connection. This makes the test runner
not get stuck, since supertest closes the server on each
request, but can't close the database connection.
- Make app.js and utils.js not log anything to the console
if in a test environment in order to keep the test runner
output intact.